Kata: Help the bookseller
Calculate the stock total based on category of books
https://www.codewars.com/kata/54dc6f5a224c26032800005c/
Level: 6kyu

Some issues with the description of this kata and test edge cases not covered in
the description of the kata.  This code does not pass all the tests on Codewars
although it does pass the tests in the project.