Added large sequence to tests for Max subarry sum kata
Whilst running all tests on CodeWars its sometimes possible to get the test data
for those tests that fail in the output window.

To help test performance, I copied one of the larger sequences used to test
solutions.  This was added to its own namespace to keep the test namespace
clean.